requirements.txt
langchain
openai
python-dotenv
pypdf2
python-docx
streamlit
docx